+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0-alpha.18] - 2026-08-26
+
 ### Changed
 
 - Show Docker host badges when multiple Docker hosts or a standalone host badge are configured; implicit hosts use `host`.